+++ R/admixcor.R
@@ -11,7 +11,8 @@ admixcor <- function(
                      vertex_refine = FALSE,
                      tol = sqrt( .Machine$double.eps ), # 1e-15
                      nstep_max = 100000,
-                     report_freq = 1000
+                     report_freq = 1000,
+                     fix_L = FALSE
                      ) {
     # process options
     Q_type <- match.arg( Q_type )
@@ -29,7 +30,7 @@ admixcor <- function(
     gamma <- gamma * n / K
     
     # initialize other variables
-    Vars <- initialize( ThetaSR, K, n, Q_type, L_type )
+    Vars <- initialize( ThetaSR, K, n, Q_type, L_type, fix_L = fix_L )
     Q0 <- Vars$Q
     L0 <- Vars$L
     R0 <- Vars$R
@@ -62,7 +63,7 @@ admixcor <- function(
         
         # apply the updates, one at the time
         R1 <- update_R( ThetaSR, Q0, L0 )
-        L1 <- update_L( ThetaSR, Q0, R1, gamma, algorithm = L_algorithm )
+        L1 <- update_L( ThetaSR, Q0, R1, gamma, algorithm = L_algorithm, fix_L = fix_L )
         Q1 <- update_Q( ThetaSR, L1, R1, delta, I, algorithm = Q_algorithm, vertex_refine = vertex_refine )
         
         # calculate step sizes, to assess convergence