+++ R/admixcor2.R
@@ -33,7 +33,7 @@ admixcor2 <- function(
     alpha <- alpha * n / K
 
     # initialize other variables
-    Vars <- initialize( ThetaSR, K, n, Q_type, L_type )
+    Vars <- initialize( ThetaSR, K, n, Q_type, L_type, fix_L = TRUE )
     Q0 <- Vars$Q
     L0 <- Vars$L
     Psi0 <- tcrossprod( L0 )